requirements.txt
django<2.2,>=1.11
psycopg2>=2.7,<3.0
django-oscar
ipython